OAK-9687 (refactor): move tracker code out of the query package
